console: Refactor sidebar context for better encapsulation
