+++ pkg/webui/console/views/app/index.js
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import React, { useCallback, useContext, useEffect, useRef, useState } from 'react'
+import React, { useCallback, useContext, useEffect, useRef } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import {
   Route,
@@ -43,7 +43,7 @@ import LogBackInModal from '@console/containers/log-back-in-modal'
 import Sidebar from '@console/containers/sidebar'
 import EventSplitFrameContext from '@console/containers/event-split-frame/context'
 import Logo from '@console/containers/logo'
-import SidebarContext from '@console/containers/sidebar/context'
+import { SidebarContextProvider } from '@console/containers/sidebar/context'
 
 import OverviewRoutes from '@console/views/overview'
 import Applications from '@console/views/applications'
@@ -105,68 +105,6 @@ const Layout = () => {
     isActive,
   } = useContext(EventSplitFrameContext)
 
-  // For the mobile side menu drawer functionality.
-  const [isDrawerOpen, setIsDrawerOpen] = useState(false)
-  const [isMinimized, setIsMinimized] = useState(false)
-  const [isSideBarHovered, setIsSideBarHovered] = useState(false)
-  const node = useRef()
-
-  const openDrawer = useCallback(() => {
-    setIsDrawerOpen(true)
-    document.body.classList.add(style.scrollLock)
-  }, [])
-
-  const closeDrawer = useCallback(() => {
-    setIsDrawerOpen(false)
-    document.body.classList.remove(style.scrollLock)
-  }, [])
-
-  const handleMouseMove = useCallback(
-    e => {
-      if (e.clientX <= 20 && isMinimized) {
-        // If the mouse is within 20px of the left edge
-        setIsDrawerOpen(true)
-      } else if (e.clientX >= 550 && isMinimized) {
-        // If the mouse is within 300px of the sidebar
-        setIsDrawerOpen(false)
-      }
-    },
-    [isMinimized],
-  )
-
-  useEffect(() => {
-    const onClickOutside = e => {
-      if (isDrawerOpen && node.current && !node.current.contains(e.target)) {
-        closeDrawer()
-      }
-    }
-
-    if (isMinimized) {
-      document.addEventListener('mousemove', handleMouseMove)
-      return () => document.removeEventListener('mousemove', handleMouseMove)
-    }
-
-    if (isDrawerOpen) {
-      document.addEventListener('mousedown', onClickOutside)
-      return () => document.removeEventListener('mousedown', onClickOutside)
-    }
-  }, [isDrawerOpen, closeDrawer, isMinimized, handleMouseMove])
-
-  // Pass this function to the header prop `onMenuClick`.
-  const onDrawerExpandClick = useCallback(() => {
-    if (!isDrawerOpen) {
-      openDrawer()
-    } else {
-      closeDrawer()
-    }
-  }, [isDrawerOpen, openDrawer, closeDrawer])
-  // End of mobile side menu drawer functionality
-
-  const onMinimizeToggle = useCallback(async () => {
-    setIsMinimized(prev => !prev)
-    setIsDrawerOpen(false)
-  }, [setIsMinimized])
-
   useEffect(() => {
     if (main.current) {
       main.current.scrollTop = 0
@@ -185,26 +123,10 @@ const Layout = () => {
           <div id="modal-container" />
           <div id="dropdown-container" className="pos-absolute-container" />
           <div className="d-flex">
-            <SidebarContext.Provider
-              value={{
-                onMinimizeToggle,
-                isMinimized,
-                setIsMinimized,
-                closeDrawer,
-                onDrawerExpandClick,
-              }}
-            >
-              <Sidebar
-                isDrawerOpen={isDrawerOpen}
-                paddingBottom={splitFrameHeight}
-                isSideBarHovered={isSideBarHovered}
-                setIsHovered={setIsSideBarHovered}
-              />
+            <SidebarContextProvider>
+              <Sidebar />
               <div className="w-full h-vh d-flex direction-column">
-                <Header
-                  onMenuClick={onDrawerExpandClick}
-                  setIsSideBarHovered={setIsSideBarHovered}
-                />
+                <Header />
                 <main
                   className={classnames(style.main, 'd-flex', 'flex-column', 'h-full', 'flex-grow')}
                   ref={main}
@@ -241,7 +163,7 @@ const Layout = () => {
                   </WithAuth>
                 </main>
               </div>
-            </SidebarContext.Provider>
+            </SidebarContextProvider>
           </div>
 
           <div id="split-frame" className={style.splitFrame} />
